+++ src/main/java/com/example/StudyFriends/repositories/FriendRep.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FriendRep extends JpaRepository<Friend,Integer> {
+public interface FriendRep extends JpaRepository<Friend,Long> {
 }
