+++ src/main/java/com/example/StudyFriends/repositories/PlayerRep.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlayerRep extends JpaRepository<Player,Integer> {
+public interface PlayerRep extends JpaRepository<Player,Long> {
 }
