+++ src/main/java/com/example/StudyFriends/services/FriendService.java
@@ -4,12 +4,9 @@ import com.example.StudyFriends.exceptions.ResourceNotFoundException;
 import com.example.StudyFriends.model.Friend;
 import com.example.StudyFriends.model.Player;
 import com.example.StudyFriends.repositories.FriendRep;
-import com.example.StudyFriends.repositories.FriendVisitRep;
 import com.example.StudyFriends.repositories.PlayerRep;
-import jakarta.persistence.criteria.CriteriaBuilder;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -21,10 +18,10 @@ public class FriendService {
     private final PlayerRep playerRep;
 
 
-    public Optional<Friend> getFriendById(Integer id){
+    public Optional<Friend> getFriendById(Long id){
         return friendRep.findById(id);
     }
-    public List<Friend> getAllFriendsOfPlayer(Integer playerId){
+    public List<Friend> getAllFriendsOfPlayer(Long playerId){
         Player player = playerRep.findById(playerId)
                 .orElseThrow(() -> new ResourceNotFoundException("Player", playerId));
         return player.getFriends();
